+++ src/OpenTelemetry.Resources.Process/ProcessDetector.cs
@@ -1,6 +1,9 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+using System.Globalization;
+using CurrentProcess = System.Diagnostics.Process;
+
 namespace OpenTelemetry.Resources.Process;
 
 /// <summary>
@@ -14,20 +17,96 @@ internal sealed class ProcessDetector : IResourceDetector
     /// <returns>Resource with key-value pairs of resource attributes.</returns>
     public Resource Detect()
     {
-        return new Resource(new List<KeyValuePair<string, object>>(2)
+        GetProcessAttributes(
+            out int processId,
+            out DateTime? creationTime,
+            out string? processPath,
+            out string? title);
+
+        var attributes = new List<KeyValuePair<string, object>>(8)
         {
             new(ProcessSemanticConventions.AttributeProcessOwner, Environment.UserName),
-#if NET
-            new(ProcessSemanticConventions.AttributeProcessPid, Environment.ProcessId),
-        });
-#else
-            new(ProcessSemanticConventions.AttributeProcessPid, GetProcessPid()),
-        });
-        static int GetProcessPid()
+            new(ProcessSemanticConventions.AttributeProcessPid, processId),
+            new(ProcessSemanticConventions.AttributeProcessWorkingDirectory, Environment.CurrentDirectory),
+        };
+
+        if (creationTime is { } startTime)
         {
-            using var process = System.Diagnostics.Process.GetCurrentProcess();
-            return process.Id;
+            attributes.Add(new(ProcessSemanticConventions.AttributeProcessCreationTime, startTime.ToString("O", CultureInfo.InvariantCulture)));
         }
+
+        if (title is { Length: > 0 })
+        {
+            attributes.Add(new(ProcessSemanticConventions.AttributeProcessTitle, title));
+        }
+
+#if NET
+        processPath ??= Environment.ProcessPath;
 #endif
+
+        // We only set the count to avoid the need to implement redaction.
+        // See https://github.com/open-telemetry/semantic-conventions/blob/v1.40.0/docs/resource/process.md#selecting-process-attributes.
+        var commandArgs = GetCommandLineArgs();
+
+        if (commandArgs.Length > 0)
+        {
+#if NET
+            processPath ??= commandArgs[0];
+#else
+            processPath = commandArgs[0];
+#endif
+
+            attributes.Add(new(ProcessSemanticConventions.AttributeProcessCommand, commandArgs[0]));
+
+            // Do not count the executable path as an argument
+            int count = commandArgs.Length - 1;
+
+            attributes.Add(new(ProcessSemanticConventions.AttributeProcessArgsCount, count));
+        }
+
+        if (!string.IsNullOrEmpty(processPath))
+        {
+            attributes.Add(new(ProcessSemanticConventions.AttributeProcessExecutablePath, processPath!));
+        }
+
+        return new Resource(attributes);
+
+        static string[] GetCommandLineArgs()
+        {
+            try
+            {
+                return Environment.GetCommandLineArgs();
+            }
+            catch (NotSupportedException)
+            {
+                return [];
+            }
+        }
+
+        static void GetProcessAttributes(
+            out int processId,
+            out DateTime? creationTime,
+            out string? processPath,
+            out string? title)
+        {
+            using var process = CurrentProcess.GetCurrentProcess();
+            processId = process.Id;
+
+            creationTime = SafeGet(process, (p) => p.StartTime);
+            processPath = SafeGet(process, (p) => p.MainModule?.FileName);
+            title = SafeGet(process, (p) => p.MainWindowTitle);
+
+            static T? SafeGet<T>(CurrentProcess process, Func<CurrentProcess, T> getter)
+            {
+                try
+                {
+                    return getter(process);
+                }
+                catch (Exception)
+                {
+                    return default;
+                }
+            }
+        }
     }
 }