+++ src/OpenTelemetry.Resources.Process/CHANGELOG.md
@@ -5,6 +5,12 @@
 * Updated OpenTelemetry core component version(s) to `1.15.1`.
   ([#4020](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4020))
 
+* Add new process resource attributes: `process.args_count`,
+  `process.command`, `process.creation.time`,
+  `process.executable.path`, `process.title` and
+  `process.working_directory`.
+  ([#4036](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4036))
+
 ## 1.15.0-beta.1
 
 Released 2026-Jan-21